# perf(dataobj): improve parallelized read spead of Zstd