Previously, a process-wide synchronous Zstd decoder was used for decompressing Zstd data. Its performance degraded as the number of goroutines decompressing Zstd data increased.

This commit removes the synchronous Zstd decoder in favor of pooled decoders that decode in parallel. Not only does this allow decoding from a single goroutine 1 GBps faster, but also makes decoding from multiple goroutines up to ~6.5x faster.

    goos: darwin
    goarch: arm64
    pkg: github.com/grafana/loki/v3/pkg/dataobj/internal/dataset
    cpu: Apple M2 Ultra
                               │    before    │                 after                 │
                               │    sec/op    │    sec/op      vs base                │
    _page_Decode/workers=1-24     17.61m ± 5%   11.73m ± 130%  -33.40% (p=0.023 n=10)
    _page_Decode/workers=2-24     34.49m ± 1%   12.46m ±   3%  -63.87% (p=0.000 n=10)
    _page_Decode/workers=5-24     85.21m ± 1%   14.42m ±   3%  -83.08% (p=0.000 n=10)
    _page_Decode/workers=10-24   170.46m ± 2%   22.60m ±   2%  -86.74% (p=0.000 n=10)
    geomean                       54.50m        14.77m         -72.89%

                               │   before    │                 after                 │
                               │   MBps/op   │   MBps/op     vs base                 │
    _page_Decode/workers=1-24    2.004k ± 4%   3.009k ± 57%   +50.15% (p=0.023 n=10)
    _page_Decode/workers=2-24    1.023k ± 1%   2.832k ±  3%  +176.78% (p=0.000 n=10)
    _page_Decode/workers=5-24     414.1 ± 1%   2447.5 ±  3%  +490.97% (p=0.000 n=10)
    _page_Decode/workers=10-24    207.1 ± 2%   1561.5 ±  2%  +654.17% (p=0.000 n=10)
    geomean                       647.5        2.389k        +268.91%

However, this change also makes memory usage seemingly appear worse, which I believe is a false positive:

    goos: darwin
    goarch: arm64
    pkg: github.com/grafana/loki/v3/pkg/dataobj/internal/dataset
    cpu: Apple M2 Ultra
                               │    before     │                     after                     │
                               │     B/op      │       B/op         vs base                    │
    _page_Decode/workers=1-24      194.0 ±  1%       2817.0 ±  44%    +1352.06% (p=0.000 n=10)
    _page_Decode/workers=2-24      524.0 ± 30%    1751949.5 ±  54%  +334241.51% (p=0.000 n=10)
    _page_Decode/workers=5-24    1.480Ki ± 36%   1308.696Ki ± 217%   +88297.43% (p=0.000 n=10)
    _page_Decode/workers=10-24   2.909Ki ± 13%   3623.577Ki ±  69%  +124477.56% (p=0.000 n=10)
    geomean                        823.1            386.5Ki          +47985.45%

                               │   before   │                after                 │
                               │ allocs/op  │  allocs/op   vs base                 │
    _page_Decode/workers=1-24    5.000 ± 0%   23.500 ± 6%  +370.00% (p=0.000 n=10)
    _page_Decode/workers=2-24    9.000 ± 0%   50.000 ± 4%  +455.56% (p=0.000 n=10)
    _page_Decode/workers=5-24    23.00 ± 9%   116.50 ± 6%  +406.52% (p=0.000 n=10)
    _page_Decode/workers=10-24   49.00 ± 4%   233.50 ± 2%  +376.53% (p=0.000 n=10)
    geomean                      15.01         75.19       +401.05%

I believe this is a false positive due to how the decoders work now: whenever a pooled decoder is released by the garbage collector, the next decoder needs to do a warmup of a high amount of allocations.